Extract system contract addresses
